Rename traits.rs to width.rs
